refactor: session 패키지 생성 및 이동
