+++ tomcat/src/main/java/org/apache/coyote/session/HttpSessionManger.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.session;
 
 import java.io.IOException;
 import java.util.HashMap;
